pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "MRI2FE"
authors = [{ name = "Turner Jennings", email = "turner.jennings@outlook.com" }]
description = "Workflows for automated FE head model generation from MRI/MRE data"
readme = "README.md"
license = { file = "LICENSE" }
version = "0.1.1"

keywords = ["MRI","FE model"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
]

requires-python = ">=3.8.1"
dependencies = ["numpy", "matplotlib","antspyx","lasso-python","scipy","julia"]

[project.optional-dependencies]
dev = ["pytest", "ruff"]


[project.urls]
Repository = "https://github.com/turnerjennings/MRI2FE.git"
Documentation = "https://turnerjennings.github.io/MRI2FE/"
Issues = "https://turnerjennings.github.io/MRI2FE/issues"

[tool.pytest.ini_options]
testpaths = "test"

[tool.ruff]
line-length = 79
indent-width = 4
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint]
fixable = ["ALL"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings.
quote-style = "double"

# Indent with spaces, rather than tabs.
indent-style = "space"

# Automatically detect the appropriate line ending.
line-ending = "auto"